refactor(docker/images): move image status types
