+++ app/react/docker/components/ImageStatus/helpers.ts
@@ -4,7 +4,7 @@ import UpdatesAvailable from '@/assets/ico/icon_updates-available.svg?c';
 import UpToDate from '@/assets/ico/icon_up-to-date.svg?c';
 import UpdatesUnknown from '@/assets/ico/icon_updates-unknown.svg?c';
 
-import { ImageStatus } from '../../images/types';
+import { ImageStatus } from './types';
 
 export function statusIcon(status: ImageStatus) {
   switch (status.Status) {
